Spike for Eclipse import of JVM Software Model based projects
It's throwaway code.
Load project with multiple components, test suites and binaries.
External dependencies are gathered on test suites only.

For example, the following build script imports:

    plugins {
        id 'jvm-component'
        id 'java-lang'
        id 'junit-test-suite'
    }
    repositories {
        jcenter()
    }
    model {
        components {
            main(JvmLibrarySpec) {
                dependencies {
                    module 'ch.qos.logback:logback-classic:1.1.3'
                }
                binaries.all {
                    sources {
                        alternate(JavaSourceSet) {
                            source.srcDir 'src/alternate/java'
                        }
                    }
                }
            }
            another(JvmLibrarySpec) {}
        }
        testSuites {
            test(JUnitTestSuiteSpec) {
                jUnitVersion '4.12'
                testing $.components.main
            }
            anotherTest(JUnitTestSuiteSpec) {
                jUnitVersion '4.12'
                testing $.components.another
            }
        }
    }

Eclipse has a single flat class path for all the sources of a given
project. So gathering dependencies on test suites only is enough for
this case.

One case that wouldn't work is a single component with external
dependencies but no attached test suite. More work needs to be done to
achieve that.