Replace usage of `Set<Integer>` and `Map<Integer, ...>` with `BitSet` and expandable array list
This commit expands on the fact we know use (small) positive component ids in dependency resolution,
allowing us to use more efficient data structures. Instead of hash sets of (long, integer) we can
now rely on a bitset. Where we had a hash map, we know have either an array, or an array list which
capacity is bounded by the maximum component id. As long as the max component id is small, this is
both more memory and CPU efficient.