Optimize configuration creation time
This commit optimizes the configuration creation time, by avoiding to create a set of mutation rules, knowing
that in the vast majority of cases there are only 1. It also optimizes iteration through `all` and `iterator`
by shortcutting if the backing store is empty, and that we can tell that it is empty in constant time.