Spike faster exclude rule merging
This commit spikes faster exclude rules merging, by:

- interning `org.gradle.api.internal.artifacts.DefaultModuleIdentifier`
- adding some "fast path" for merging
- introducing caches in `ModuleExclusions`

This is just spiking how fast we can go with that approach. With the customer's test project, merging is not a hotspot
anymore. But the caches we introduces are static, which is no good. So if we go that path, we should turn `ModuleExclusions`
into a build scoped service, which will require more work.