Introduce an API to make dynamic version listing faster
This commit introduces a public API to support custom dynamic version resolvers. This can be attached to a repository, in order to make resolution
faster in case the user knows better how to provide the latest version. This is typically the case when Ivy repositories are in use: since no
metadata file is available, Gradle has to scan all versions until it finds the latest release (or integration version, ...) which is very inefficient.
With this API, a user can implement any arbitrary version listing strategy: using a remote file, or something else.

This API uses a class literal, which means the responsibility of creating a lister is delegated to Gradle. This also means that the potential caching
would have to be implemented on the Gradle side and cannot be provided by the user.