Prevent the same module from being resolved concurrently
There are 2 different guards in place: the first one is used during resolution of a module, and excludes the version
number from the guard. This is done so that we make sure that `foo:bar:1+` and `foo:bar:latest.release` are not
resolved concurrently. The second guard is on artifacts, and for this, we can use the full component identifier as
the key. This means that the locking for artifacts is finer grained that the one for modules.